adaptations to support Kibana 7.8.0
