+++ public/components/document_table_vis_data.tsx
@@ -22,8 +22,8 @@ import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n/react';
 import { EuiButtonEmpty, EuiDragDropContext, euiDragDropReorder, EuiDroppable, EuiFlexGroup, EuiFlexItem, EuiFormErrorText, EuiPanel, EuiSpacer, EuiTitle } from '@elastic/eui';
 
-import { VisOptionsProps } from '../../../../src/legacy/core_plugins/vis_default_editor/public';
-import { NumberInputOption, SelectOption } from '../../../../src/legacy/core_plugins/vis_type_vislib/public';
+import { VisOptionsProps } from '../../../../src/plugins/vis_default_editor/public';
+import { NumberInputOption, SelectOption } from '../../../../src/plugins/charts/public';
 import { FieldColumnEditor } from './field_column';
 import { FieldParamEditor } from './field';
