+++ public/data_load/enhanced-table-request-handler.js
@@ -23,12 +23,16 @@ import { SearchSource } from '../../../../src/plugins/data/public/search/search_
 import { RequestAdapter, DataAdapter } from '../../../../src/plugins/inspector/public';
 import { getSearchService, getQueryService } from '../../../../src/plugins/data/public/services';
 import { serializeAggConfig } from '../../../../src/plugins/data/public/search/expressions/utils';
+import { npStart } from '../legacy_imports';
+import { search } from '../../../../src/plugins/data/public'
 
 export async function enhancedTableRequestHandler ({
   partialRows,
   metricsAtAllLevels,
   visParams,
   timeRange,
+  timeFields,
+  indexPattern,
   query,
   filters,
   inspectorAdapters,
@@ -42,7 +46,11 @@ export async function enhancedTableRequestHandler ({
   const aggs = searchService.aggs.createAggConfigs(index, aggConfigsState);
 
   // create search source with query parameters
-  const searchSource = new SearchSource();
+
+  let searchSourceDeps = npStart.core;
+  searchSourceDeps.search = searchService.search
+
+  const searchSource = new SearchSource({}, searchSourceDeps);
   searchSource.setField('index', aggs.indexPattern);
   const hitsSize = (visParams.hitsSize !== undefined ? visParams.hitsSize : 0);
   searchSource.setField('size', hitsSize);
