--- public/enhanced-table-vis.js
@@ -1,135 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { i18n } from '@kbn/i18n';
-import { AggGroupNames } from '../../../src/plugins/data/public';
-import { Schemas } from '../../../src/legacy/core_plugins/vis_default_editor/public';
-import { prepareJson, prepareString } from '../../../src/legacy/core_plugins/visualizations/public/np_ready/public/legacy/build_pipeline';
-
-import tableVisTemplate from './enhanced-table-vis.html';
-import { EnhancedTableVisualizationController } from './vis_controller';
-import { enhancedTableRequestHandler } from './data_load/enhanced-table-request-handler';
-import { enhancedTableResponseHandler } from './data_load/enhanced-table-response-handler';
-import { EnhancedTableOptions } from './components/enhanced_table_vis_options';
-
-
-// define the visType object, which kibana will use to display and configure new Vis object of this type.
-export const enhancedTableVisTypeDefinition = {
-  type: 'table',
-  name: 'enhanced-table',
-  title: i18n.translate('visTypeEnhancedTable.visTitle', {
-    defaultMessage: 'Enhanced Table'
-  }),
-  icon: 'visTable',
-  description: i18n.translate('visTypeDocumentTable.visDescription', {
-    defaultMessage: 'Same functionality than Data Table, but with enhanced features like computed columns, filter bar and pivot table.'
-  }),
-  visualization: EnhancedTableVisualizationController,
-  visConfig: {
-    defaults: {
-      perPage: 10,
-      showPartialRows: false,
-      showMetricsAtAllLevels: false,
-      sort: {
-        columnIndex: null,
-        direction: null
-      },
-      showTotal: false,
-      totalFunc: 'sum',
-      computedColumns: [],
-      computedColsPerSplitCol: false,
-      hideExportLinks: false,
-      stripedRows: false,
-      showFilterBar: false,
-      filterCaseSensitive: false,
-      filterBarHideable: false,
-      filterAsYouType: false,
-      filterTermsSeparately: false,
-      filterHighlightResults: false,
-      filterBarWidth: '25%'
-    },
-    template: tableVisTemplate
-  },
-  editorConfig: {
-    optionsTemplate: EnhancedTableOptions,
-    schemas: new Schemas([
-      {
-        group: AggGroupNames.Metrics,
-        name: 'metric',
-        title: i18n.translate('visTypeTable.tableVisEditorConfig.schemas.metricTitle', {
-          defaultMessage: 'Metric'
-        }),
-        aggFilter: ['!geo_centroid', '!geo_bounds'],
-        aggSettings: {
-          top_hits: {
-            allowStrings: true
-          }
-        },
-        min: 1,
-        defaults: [{ type: 'count', schema: 'metric' }]
-      },
-      {
-        group: AggGroupNames.Buckets,
-        name: 'split',
-        title: i18n.translate('visTypeTable.tableVisEditorConfig.schemas.splitTitle', {
-          defaultMessage: 'Split table'
-        }),
-        min: 0,
-        max: 1,
-        aggFilter: ['!filter']
-      },
-      {
-        group: AggGroupNames.Buckets,
-        name: 'bucket',
-        title: i18n.translate('visTypeTable.tableVisEditorConfig.schemas.bucketTitle', {
-          defaultMessage: 'Split rows'
-        }),
-        aggFilter: ['!filter']
-      },
-      {
-        group: AggGroupNames.Buckets,
-        name: 'splitcols',
-        title: i18n.translate('tableVis.tableVisEditorConfig.schemas.splitcolsTitle', {
-          defaultMessage: 'Split cols'
-        }),
-        aggFilter: ['!filter'],
-        max: 1,
-        editor: '<div class="hintbox"><i class="fa fa-danger text-info"></i> This bucket must be the last one</div>'
-      }
-    ])
-  },
-  requestHandler: enhancedTableRequestHandler,
-  responseHandler: enhancedTableResponseHandler,
-  hierarchicalData: function (vis) {
-    return Boolean(vis.params.showPartialRows || vis.params.showMetricsAtAllLevels);
-  },
-  toExpression: function (vis) {
-    const visConfig = { ...vis.params };
-    const { indexPattern, aggs } = vis.data;
-    let pipeline = `enhanced_table_visualization type='${vis.type.name}'
-      ${prepareJson('visConfig', visConfig)}
-      ${prepareJson('aggConfigs', aggs.aggs)}
-      metricsAtAllLevels=${vis.isHierarchical()}
-      partialRows=${vis.type.requiresPartialRows || vis.params.showPartialRows || false} `;
-    if (indexPattern) {
-      pipeline += `${prepareString('index', indexPattern.id)}`;
-    }
-    return pipeline;
-  }
-};
