+++ public/index.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { PluginInitializerContext } from '../../../src/core/public';
+import { PluginInitializerContext } from 'kibana/public';
 import { EnhancedTablePlugin as Plugin } from './plugin';
 
 export function plugin(initializerContext: PluginInitializerContext) {
